Add rudimentary support for C quoted embedding
The conversion from C types to Haskell types is severly incomplete (see
`cTypeToHsType`).  I also have to think about what to do with
`typedef`s.

Right now the syntax is just

    int        (int x, int y)                      { x + y }
    ^          ^                                   ^
    ret type   parameters capturing Haskell vars   C code using the parameters

I plan to also extend it so that if `x_int` is present in the C code, it
is automatically added to the list of parameters, so that the above will
be expressible as

    int { x_int + y_int }

or similar.

If a variable appears "declared" multiple times, the types must
match. So that this is OK

    int (int x) { x_int + y_int } ==> int (int x, int y) { x + y }

this is also OK

    int { x_int + x_int } ==> int (int x) { x + y }

so is this

    int { x_int + x} ==> int (int x) { x + x }

but this is not

    int (int x) { x_double + y_int } ==> ?