Improve tests and parsing.
Now parsing is done entirely piggybacking the type parser for C.  This
is because the type parser can parse proper type declarations.  So
basically the parser for typed expressions/items now expects a (possibly
anonymous) function declarations.  If it's not a function declaration it
just leaves the type as is and uses it as the return type, with no parameters.