Support abstract declarations.
Right now one has to explicitely parse abstract declarations.  This is
because we cannot distinguish between abstract and non-abstract
declarations if we don't know what is a type and what isn't.

Even in simple examples, such as `int foo`, if we can't make a
commitment on whether `foo` is a type or not, the parse result will
change depending on whether we interpret the declaration as abstract or
not.  If not abstract, it will be parsed as "`foo` is an int", if
abstract, we'll have "something of type `int foo`".

I think that parsing always non-abstractly is the right choice, since it
is the common practice in C.  Moreover, in our use-case for inline-c
we *always* need the identifier name since we use declarations to
capture variables, with one exception: function prototypes, where the
identifier is not required.  I think it's a pretty small price to
pay (one can always manually insert dummy `_` identifiers).

However, we still want support for abstract declarations explicitly so
that we can have abstract "typed" C blocks such as

   int (int x, int y) { ... }