Move the unsafe quasiquote variants into separate module.
The main reason for this is that it affords the user choice as to how
these quasi-quoters should be named. Dependending on the use case, some
users will want to be using the unasfe variants everywhere, only seldom
the safe ones. Or vice versa. Requiring the user to write '_unsafe'
everywhere can become pretty verbose. If the unsafe quasi-quotes have
the same name but are in a different module, the user can choose the
name of the qualified import, or in fact choose not to use unsafe
quasi-quoter at all.

The second reason is to unclutter the API of the main module.