Use file names rather than module names.
Since we're pairing Haskell files with C files, it makes more sense to
use the file name.  I'm not sure if it's possible to compile two modules
with the same name in one run of GHC, but in this way we make sure that
we don't run into trouble even it is.